Fix/010 add csr csc sorted return (#31)
* - Krylov Power iteration and Inverse Power integration.

* - Bug fix.

* - added test cases for edge cases.

* - `SparseLinearOp` validation module (`sparse::validate`) for checking the linear operator trait bounds.

* `ConvergenceError::NonFinite` in `src/krylov/mod.rs`, with a doc-comment explaining it as an explicit exception to the
  crate's general NaN/Inf-propagation policy — bounded compute budget on embedded targets
  shouldn't be spent iterating on poisoned state.

* Fix:
- add Result-based error handling to sparse matrix operations, validate dimension overflows with checked_add and try_from
 Docs:
 - add # Errors sections to doc comments, update examples with .unwrap()
 Tests: update all tests for Result
  handling, add dimension overflow edge case tests

* Fix #10: add_csr/add_csc now return SortedCsrMatrix/SortedCscMatrix

Both addition algorithms already produced sorted, deduplicated indices per
row/column, so the return type is changed to reflect that guarantee instead
of the plain, unsorted-by-contract Csr/CscMatrix. Adds a SortedCscMatrix
counterpart to the existing SortedCsrMatrix.


* Fix Scalar::sin/cos precision test by adding quadrant range reduction

The precision test over [-2*pi, 2*pi] was failing: sin/cos ran a fixed
20-term Taylor series directly on the input angle, which only converges
tightly near zero and lost precision far from it, especially near zero
crossings.

Reduces the input to the nearest quadrant (|remainder| <= pi/4) via a
Cody-Waite hi/lo split before running the Taylor series, so the series
always operates on a small argument. f32's reduction is done in f64
arithmetic (then narrowed) since f32 doesn't have enough bits to represent
the pi/2 split precisely enough near a zero crossing.


* - Fix CHANGELOG.md for readability.

---------

+++ src/sparse/coo.rs
@@ -43,8 +43,8 @@ pub enum CooError {
 pub struct CooMatrix<T> {
     rows: usize,
     cols: usize,
-    row_indices: Vec<usize>,
-    col_indices: Vec<usize>,
+    row_indices: Vec<u32>,
+    col_indices: Vec<u32>,
     values: Vec<T>,
 }
 
@@ -52,8 +52,8 @@ impl<T> CooMatrix<T> {
     pub(super) fn new_raw(
         rows: usize,
         cols: usize,
-        row_indices: Vec<usize>,
-        col_indices: Vec<usize>,
+        row_indices: Vec<u32>,
+        col_indices: Vec<u32>,
         values: Vec<T>,
     ) -> Self {
         Self {
@@ -65,7 +65,7 @@ impl<T> CooMatrix<T> {
         }
     }
 
-    pub(super) fn into_raw_parts(self) -> (usize, usize, Vec<usize>, Vec<usize>, Vec<T>) {
+    pub(super) fn into_raw_parts(self) -> (usize, usize, Vec<u32>, Vec<u32>, Vec<T>) {
         (
             self.rows,
             self.cols,
@@ -98,20 +98,20 @@ impl<T> CooMatrix<T> {
     pub fn new(
         rows: usize,
         cols: usize,
-        row_indices: Vec<usize>,
-        col_indices: Vec<usize>,
+        row_indices: Vec<u32>,
+        col_indices: Vec<u32>,
         values: Vec<T>,
     ) -> Result<Self, CooError> {
         if row_indices.len() != col_indices.len() || row_indices.len() != values.len() {
             return Err(CooError::LengthMismatch);
         }
         for &r in &row_indices {
-            if r >= rows {
+            if r as usize >= rows {
                 return Err(CooError::RowIndexOutOfBounds);
             }
         }
         for &c in &col_indices {
-            if c >= cols {
+            if c as usize >= cols {
                 return Err(CooError::ColIndexOutOfBounds);
             }
         }
@@ -177,7 +177,7 @@ impl<T> CooMatrix<T> {
     /// let m = CooMatrix::new(2, 2, vec![0, 1], vec![1, 0], vec![5.0_f64, 9.0]).unwrap();
     /// assert_eq!(m.row_indices(), &[0, 1]);
     /// ```
-    pub fn row_indices(&self) -> &[usize] {
+    pub fn row_indices(&self) -> &[u32] {
         &self.row_indices
     }
 
@@ -191,7 +191,7 @@ impl<T> CooMatrix<T> {
     /// let m = CooMatrix::new(2, 2, vec![0, 1], vec![1, 0], vec![5.0_f64, 9.0]).unwrap();
     /// assert_eq!(m.col_indices(), &[1, 0]);
     /// ```
-    pub fn col_indices(&self) -> &[usize] {
+    pub fn col_indices(&self) -> &[u32] {
         &self.col_indices
     }
